+++ custom_components/supernotify/hass_api.py
@@ -5,6 +5,7 @@ from dataclasses import dataclass
 from functools import partial
 from typing import TYPE_CHECKING, Any
 
+from homeassistant.components.person import ATTR_USER_ID
 from homeassistant.const import CONF_ACTION, CONF_DEVICE_ID
 from homeassistant.helpers.aiohttp_client import async_get_clientsession
 from homeassistant.helpers.event import async_track_state_change_event, async_track_time_change
@@ -67,19 +68,20 @@ _LOGGER = logging.getLogger(__name__)
 
 
 @dataclass
-class MobileAppInfo:
-    mobile_app_id: str
-    device_name: str | None
-    device_id: str | None
-    device_tracker: str | None
-    action: str | None
-    user_id: str | None
-    manufacturer: str | None
-    model: str | None
-    os_name: str | None
-    os_version: str | None
-    app_version: str | None
+class DeviceInfo:
+    device_id: str
     device_labels: list[str]
+    mobile_app_id: str | None = None
+    device_name: str | None = None
+    device_tracker: str | None = None
+    action: str | None = None
+    user_id: str | None = None
+    manufacturer: str | None = None
+    model: str | None = None
+    os_name: str | None = None
+    os_version: str | None = None
+    app_version: str | None = None
+    identifiers: set[tuple[str, str]] | None = None
 
     def as_dict(self) -> dict[str, str | list[str] | None]:
         return {
@@ -113,10 +115,10 @@ class HomeAssistantAPI:
         self._device_registry: dr.DeviceRegistry | None = None
         self._service_info: dict[tuple[str, str], Any] = {}
         self.unsubscribes: list[CALLBACK_TYPE] = []
-        self.mobile_apps_by_tracker: dict[str, MobileAppInfo] = {}
-        self.mobile_apps_by_app_id: dict[str, MobileAppInfo] = {}
-        self.mobile_apps_by_device_id: dict[str, MobileAppInfo] = {}
-        self.mobile_apps_by_user_id: dict[str, list[MobileAppInfo]] = {}
+        self.mobile_apps_by_tracker: dict[str, DeviceInfo] = {}
+        self.mobile_apps_by_app_id: dict[str, DeviceInfo] = {}
+        self.mobile_apps_by_device_id: dict[str, DeviceInfo] = {}
+        self.mobile_apps_by_user_id: dict[str, list[DeviceInfo]] = {}
 
     def initialize(self) -> None:
         self.hass_name = self._hass.config.location_name
@@ -288,6 +290,7 @@ class HomeAssistantAPI:
         this_trace: ActionTrace | None = None
         if DATA_TRACE not in self._hass.data:
             _LOGGER.warning("SUPERNOTIFY tracing not configured, attempting to set up")
+
             await homeassistant.components.trace.async_setup(self._hass, {})  # type: ignore
         with trace_action(self._hass, trace_name or "anon_condition") as cond_trace:
             cond_trace.set_trace(trace_get())
@@ -342,7 +345,7 @@ class HomeAssistantAPI:
     ) -> bool | None:
         try:
             if not condition_variables:
-                _LOGGER.warning("SUPERNOTIFY No cond vars provided for condition: %s", conditions)
+                _LOGGER.warning("SUPERNOTIFY No cond vars provided for condition")
             return conditions(condition_variables.as_dict() if condition_variables is not None else None)
         except Exception as e:
             _LOGGER.error("SUPERNOTIFY Condition eval failed: %s", e)
@@ -378,16 +381,16 @@ class HomeAssistantAPI:
             is_fixable=is_fixable,
         )
 
-    def mobile_app_by_tracker(self, device_tracker: str) -> MobileAppInfo | None:
+    def mobile_app_by_tracker(self, device_tracker: str) -> DeviceInfo | None:
         return self.mobile_apps_by_tracker.get(device_tracker)
 
-    def mobile_app_by_id(self, mobile_app_id: str) -> MobileAppInfo | None:
+    def mobile_app_by_id(self, mobile_app_id: str) -> DeviceInfo | None:
         return self.mobile_apps_by_app_id.get(mobile_app_id)
 
-    def mobile_app_by_device_id(self, device_id: str) -> MobileAppInfo | None:
+    def mobile_app_by_device_id(self, device_id: str) -> DeviceInfo | None:
         return self.mobile_apps_by_device_id.get(device_id)
 
-    def mobile_app_by_user_id(self, user_id: str) -> list[MobileAppInfo] | None:
+    def mobile_app_by_user_id(self, user_id: str) -> list[DeviceInfo] | None:
         return self.mobile_apps_by_user_id.get(user_id)
 
     def build_mobile_app_cache(self) -> None:
@@ -399,69 +402,59 @@ class HomeAssistantAPI:
 
         found: int = 0
         complete: int = 0
-        for device in self.discover_devices("mobile_app"):
+        for mobile_app_info in self.discover_devices("mobile_app"):
             try:
-                mobile_app_id: str = f"mobile_app_{slugify(device.name)}"
+                mobile_app_id: str = f"mobile_app_{slugify(mobile_app_info.device_name)}"
                 device_tracker: str | None = None
                 notify_action: str | None = None
-                os_name: str | None = None
-                os_version: str | None = None
-                app_version: str | None = None
-                user_id: str | None = None
                 if self.has_service("notify", mobile_app_id):
                     notify_action = f"notify.{mobile_app_id}"
                 else:
                     _LOGGER.warning("SUPERNOTIFY Unable to find notify action <%s>", mobile_app_id)
-                registry_entries = ent_reg.entities.get_entries_for_device_id(device.id)
+
+                registry_entries = ent_reg.entities.get_entries_for_device_id(mobile_app_info.device_id)
                 for reg_entry in registry_entries:
                     if reg_entry.platform == "mobile_app" and reg_entry.domain == "device_tracker":
                         device_tracker = reg_entry.entity_id
-                for config_entry_id in device.config_entries:
-                    config_entry = self._hass.config_entries.async_get_entry(config_entry_id)
-                    if config_entry and config_entry.data:
-                        os_name = config_entry.data.get(ATTR_OS_NAME) or os_name
-                        os_version = config_entry.data.get(ATTR_OS_VERSION) or os_version
-                        user_id = config_entry.data.get(CONF_USER_ID) or user_id
-                        app_version = config_entry.data.get(ATTR_APP_VERSION) or app_version
 
                 if device_tracker and notify_action:
                     complete += 1
 
-                mobile_app_info: MobileAppInfo = MobileAppInfo(
-                    manufacturer=device.manufacturer,
-                    model=device.model,
-                    mobile_app_id=mobile_app_id,
-                    device_tracker=device_tracker,
-                    action=notify_action,
-                    device_id=device.id,
-                    user_id=user_id,
-                    os_name=os_name,
-                    os_version=os_version,
-                    app_version=app_version,
-                    device_name=device.name,
-                    device_labels=list(device.labels) if device.labels else [],
-                )
+                mobile_app_info.mobile_app_id = mobile_app_id
+                mobile_app_info.device_tracker = device_tracker
+                mobile_app_info.action = notify_action
+
                 found += 1
                 self.mobile_apps_by_app_id[mobile_app_id] = mobile_app_info
-                self.mobile_apps_by_device_id[device.id] = mobile_app_info
+                self.mobile_apps_by_device_id[mobile_app_info.device_id] = mobile_app_info
                 if device_tracker:
                     self.mobile_apps_by_tracker[device_tracker] = mobile_app_info
-                if user_id:
-                    self.mobile_apps_by_user_id.setdefault(user_id, [])
-                    self.mobile_apps_by_user_id[user_id].append(mobile_app_info)
+                if mobile_app_info.user_id is not None:
+                    self.mobile_apps_by_user_id.setdefault(mobile_app_info.user_id, [])
+                    self.mobile_apps_by_user_id[mobile_app_info.user_id].append(mobile_app_info)
 
             except Exception as e:
-                _LOGGER.error("SUPERNOTIFY Failure examining device %s: %s", device, e)
+                _LOGGER.error("SUPERNOTIFY Failure examining device %s: %s", mobile_app_info, e)
 
         _LOGGER.info(f"SUPERNOTIFY Found {found} enabled mobile app devices, {complete} complete config")
+
+    def device_config_info(self, device: DeviceEntry) -> dict[str, str | None]:
+        results: dict[str, str | None] = {ATTR_OS_NAME: None, ATTR_OS_VERSION: None, CONF_USER_ID: None, ATTR_APP_VERSION: None}
+        for config_entry_id in device.config_entries:
+            config_entry = self._hass.config_entries.async_get_entry(config_entry_id)
+            if config_entry and config_entry.data:
+                for attr in results:
+                    results[attr] = config_entry.data.get(attr) or results[attr]
+        return results
 
     def discover_devices(
         self,
         discover_domain: str,
         device_model_select: SelectionRule | None = None,
         device_manufacturer_select: SelectionRule | None = None,
-    ) -> list[DeviceEntry]:
-        devices: list[DeviceEntry] = []
+        device_os_select: SelectionRule | None = None,
+    ) -> list[DeviceInfo]:
+        devices: list[DeviceInfo] = []
         dev_reg: DeviceRegistry | None = self.device_registry()
         if dev_reg is None or not hasattr(dev_reg, "devices"):
             _LOGGER.warning(f"SUPERNOTIFY Unable to discover devices for {discover_domain} - no device registry found")
@@ -487,7 +480,27 @@ class HomeAssistantAPI:
                             _LOGGER.debug("SUPERNOTIFY Skipped dev %s, no manufacturer %s match", dev.name, dev.manufacturer)
                             skipped_devs += 1
                             continue
-                        devices.append(dev)
+                        device_config_info = self.device_config_info(dev)
+                        if device_os_select is not None and not device_os_select.match(device_config_info[ATTR_OS_NAME]):
+                            _LOGGER.debug(
+                                "SUPERNOTIFY Skipped dev %s, no OS %s match", dev.name, device_config_info[ATTR_OS_NAME]
+                            )
+                            skipped_devs += 1
+                            continue
+                        devices.append(
+                            DeviceInfo(
+                                device_id=dev.id,
+                                device_name=dev.name,
+                                manufacturer=dev.manufacturer,
+                                model=dev.model,
+                                user_id=device_config_info[ATTR_USER_ID],
+                                os_name=device_config_info[ATTR_OS_NAME],
+                                os_version=device_config_info[ATTR_OS_VERSION],
+                                app_version=device_config_info[ATTR_APP_VERSION],
+                                device_labels=list(dev.labels) if dev.labels else [],
+                                identifiers=dev.identifiers,
+                            )
+                        )
 
                     elif identifier:
                         # HomeKit has triples for identifiers, other domains may behave similarly