+++ custom_components/supernotify/transports/tts.py
@@ -32,7 +32,7 @@ from custom_components.supernotify.model import (
 from custom_components.supernotify.transport import Transport
 
 if TYPE_CHECKING:
-    from custom_components.supernotify.hass_api import MobileAppInfo
+    from custom_components.supernotify.hass_api import DeviceInfo
 
 _LOGGER = logging.getLogger(__name__)
 RE_VALID_MEDIA_PLAYER = r"media_player\.[A-Za-z0-9_]+"
@@ -121,7 +121,7 @@ class TTSTransport(Transport):
         at_least_one: bool = False
         for target in targets:
             bare_target = target.replace("notify.", "", 1) if target.startswith("notify.") else target
-            mobile_info: MobileAppInfo | None = self.context.hass_api.mobile_app_by_id(bare_target)
+            mobile_info: DeviceInfo | None = self.context.hass_api.mobile_app_by_id(bare_target)
             if not mobile_info or mobile_info.manufacturer == "Apple":
                 _LOGGER.debug("SUPERNOTIFY Skipping tts target that isn't confirmed as android: %s", mobile_info)
             else: