+++ tests/components/supernotify/test_delivery.py
@@ -1,7 +1,6 @@
 from unittest.mock import AsyncMock, Mock
 
 from homeassistant.const import CONF_ACTION, CONF_CONDITION
-from homeassistant.helpers.device_registry import DeviceEntry
 
 from custom_components.supernotify import (
     CONF_DELIVERY_DEFAULTS,
@@ -13,6 +12,7 @@ from custom_components.supernotify import (
 )
 from custom_components.supernotify.context import Context
 from custom_components.supernotify.delivery import Delivery
+from custom_components.supernotify.hass_api import DeviceInfo
 from custom_components.supernotify.model import Target
 from custom_components.supernotify.transports.generic import GenericTransport
 from custom_components.supernotify.transports.notify_entity import NotifyEntityTransport
@@ -92,9 +92,9 @@ def test_device_discovery(unmocked_config: Context) -> None:
         transport=GenericTransport(unmocked_config, {CONF_DEVICE_DOMAIN: ["unit_testing"], CONF_DEVICE_DISCOVERY: True}),
     )
 
-    dev: DeviceEntry = Mock(spec=DeviceEntry, id="11112222ffffeeee00009999ddddcccc")
+    dev: DeviceInfo = Mock(spec=DeviceInfo, device_id="11112222ffffeeee00009999ddddcccc")
     unmocked_config.hass_api.discover_devices = Mock(  # type: ignore
         return_value=[dev]
     )
     uut.discover_devices(unmocked_config)
-    assert uut.target.device_ids == [dev.id]  # type: ignore
+    assert uut.target.device_ids == [dev.device_id]  # type: ignore